+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/planning/BooleanNormalizer.java
@@ -27,6 +27,7 @@ import com.apple.foundationdb.record.query.expressions.AndComponent;
 import com.apple.foundationdb.record.query.expressions.BooleanComponent;
 import com.apple.foundationdb.record.query.expressions.ComponentWithChildren;
 import com.apple.foundationdb.record.query.expressions.ComponentWithSingleChild;
+import com.apple.foundationdb.record.query.expressions.NestedField;
 import com.apple.foundationdb.record.query.expressions.NotComponent;
 import com.apple.foundationdb.record.query.expressions.OrComponent;
 import com.apple.foundationdb.record.query.expressions.Query;
@@ -35,10 +36,13 @@ import com.apple.foundationdb.record.query.plan.RecordQueryPlannerConfiguration;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.Deque;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -159,7 +163,7 @@ public class BooleanNormalizer {
                 return predicate;
             }
         } else {
-            final List<List<QueryComponent>> orOfAnd = toDNF(predicate, false);
+            final List<List<QueryComponent>> orOfAnd = toDNF(predicate, false, new ArrayDeque<>());
             if (checkForDuplicateConditions) {
                 removeDuplicateConditions(orOfAnd);
             }
@@ -168,10 +172,35 @@ public class BooleanNormalizer {
     }
 
     private boolean needsNormalize(@Nullable final QueryComponent predicate) {
-        return isBooleanPredicate(predicate) &&
-               (predicate instanceof ComponentWithChildren ?
-                ((ComponentWithChildren)predicate).getChildren().stream().anyMatch(this::isBooleanPredicate) :
-                isBooleanPredicate(((ComponentWithSingleChild)predicate).getChild()));
+        // We need to normalize the expression if there is a BooleanComponent on top of another BooleanComponent,
+        // e.g., a Not on top of multiple ORs.
+        return predicate != null && depthAtLeast(predicate, 2, 0);
+    }
+
+    /**
+     * Return whether the depth of the query component tree is greater than or equal to the target depth.
+     * This only calculates the depth for {@link QueryComponent}s that the expression knows it can normalize.
+     * If a new type is added to {@link #toDNF(QueryComponent, boolean, Deque)}, this should be updated to handle
+     * the expression.
+     *
+     * @param predicate the predicate to check the Boolean depth of
+     * @param target the target depth
+     * @param booleanDepthSoFar the depth of the current {@code predicate} within a larger tree
+     * @return whether the depth meets or exceeds the target
+     */
+    private boolean depthAtLeast(@Nonnull final QueryComponent predicate, final int target, int booleanDepthSoFar) {
+        if (predicate instanceof BooleanComponent || predicate instanceof NestedField) {
+            int newDepth = booleanDepthSoFar + 1;
+            if (newDepth >= target) {
+                return true;
+            }
+            if (predicate instanceof ComponentWithChildren) {
+                return ((ComponentWithChildren)predicate).getChildren().stream().anyMatch(p -> depthAtLeast(p, target, newDepth));
+            } else if (predicate instanceof ComponentWithSingleChild) {
+                return depthAtLeast(((ComponentWithSingleChild)predicate).getChild(), target, newDepth);
+            }
+        }
+        return false;
     }
 
     private boolean shouldNormalize(@Nullable final QueryComponent predicate) {
@@ -181,10 +210,6 @@ public class BooleanNormalizer {
             // Our computation caused an integer overflow so the DNF is _definitely_ too big.
             return false;
         }
-    }
-
-    private boolean isBooleanPredicate(@Nullable final QueryComponent predicate) {
-        return predicate instanceof BooleanComponent;
     }
 
     int getNormalizedSize(@Nullable final QueryComponent predicate) {
@@ -203,6 +228,8 @@ public class BooleanNormalizer {
             return negate ? andToDNFSize(children, true) : orToDNFSize(children, false);
         } else if (predicate instanceof NotComponent) {
             return toDNFSize(((NotComponent)predicate).getChild(), !negate);
+        } else if (predicate instanceof NestedField) {
+            return toDNFSize(((NestedField)predicate).getChild(), negate);
         } else {
             return 1;
         }
@@ -241,17 +268,24 @@ public class BooleanNormalizer {
      * @return a list (to be Or'ed) of lists (to be And'ed)
      */
     @Nonnull
-    private List<List<QueryComponent>> toDNF(@Nonnull final QueryComponent predicate, final boolean negate) {
+    private List<List<QueryComponent>> toDNF(@Nonnull final QueryComponent predicate, final boolean negate, @Nonnull final Deque<String> parentPath) {
         if (predicate instanceof AndComponent) {
             final List<QueryComponent> children = ((AndComponent)predicate).getChildren();
-            return negate ? orToDNF(children, true) : andToDNF(children, false);
+            return negate ? orToDNF(children, true, parentPath) : andToDNF(children, false, parentPath);
         } else if (predicate instanceof OrComponent) {
             final List<QueryComponent> children = ((OrComponent)predicate).getChildren();
-            return negate ? andToDNF(children, true) : orToDNF(children, false);
+            return negate ? andToDNF(children, true, parentPath) : orToDNF(children, false, parentPath);
         } else if (predicate instanceof NotComponent) {
-            return toDNF(((NotComponent)predicate).getChild(), !negate);
+            return toDNF(((NotComponent)predicate).getChild(), !negate, parentPath);
+        } else if (predicate instanceof NestedField) {
+            return nestedFieldToDNF((NestedField)predicate, negate, parentPath);
         } else {
-            return Collections.singletonList(Collections.singletonList(negate ? Query.not(predicate) : predicate));
+            QueryComponent predicateForNormalization = predicate;
+            final Iterator<String> pathIterator = parentPath.descendingIterator();
+            while (pathIterator.hasNext()) {
+                predicateForNormalization = new NestedField(pathIterator.next(), predicateForNormalization);
+            }
+            return Collections.singletonList(Collections.singletonList(negate ? Query.not(predicateForNormalization) : predicateForNormalization));
         }
     }
 
@@ -262,9 +296,9 @@ public class BooleanNormalizer {
      * @return a list (to be Or'ed) of lists (to be And'ed)
      */
     @Nonnull
-    private List<List<QueryComponent>> orToDNF(@Nonnull final List<QueryComponent> children, final boolean negate) {
+    private List<List<QueryComponent>> orToDNF(@Nonnull final List<QueryComponent> children, final boolean negate, @Nonnull final Deque<String> parentPath) {
         final List<List<QueryComponent>> result = new ArrayList<>();
-        children.stream().map(p -> toDNF(p, negate)).forEach(result::addAll);
+        children.stream().map(p -> toDNF(p, negate, parentPath)).forEach(result::addAll);
         return result;
     }
 
@@ -275,24 +309,35 @@ public class BooleanNormalizer {
      * @return a list (to be Or'ed) of lists (to be And'ed)
      */
     @Nonnull
-    private List<List<QueryComponent>> andToDNF(@Nonnull final List<QueryComponent> children, final boolean negate) {
-        return andToDNF(children, 0, negate, Collections.singletonList(Collections.emptyList()));
+    private List<List<QueryComponent>> andToDNF(@Nonnull final List<QueryComponent> children, final boolean negate, @Nonnull final Deque<String> parentPath) {
+        return andToDNF(children, 0, negate, parentPath, Collections.singletonList(Collections.emptyList()));
     }
 
     @Nonnull
     private List<List<QueryComponent>> andToDNF(@Nonnull final List<QueryComponent> children, int index,
-                                                         final boolean negate,
-                                                         @Nonnull final List<List<QueryComponent>> crossProductSoFar) {
+                                                final boolean negate,
+                                                @Nonnull final Deque<String> parentPath,
+                                                @Nonnull final List<List<QueryComponent>> crossProductSoFar) {
         if (index >= children.size()) {
             return crossProductSoFar;
         }
-        return andToDNF(children, index + 1, negate,
+        return andToDNF(children, index + 1, negate, parentPath,
                 // Add each of the next child's alternatives to the each of the elements of the cross product so far.
-                toDNF(children.get(index), negate).stream().flatMap(right -> crossProductSoFar.stream().map(left -> {
+                toDNF(children.get(index), negate, parentPath).stream().flatMap(right -> crossProductSoFar.stream().map(left -> {
                     final List<QueryComponent> combined = new ArrayList<>(left);
                     combined.addAll(right);
                     return combined;
                 })).collect(Collectors.toList()));
+    }
+
+    @Nonnull
+    private List<List<QueryComponent>> nestedFieldToDNF(@Nonnull final NestedField nestedField, final boolean negate, @Nonnull final Deque<String> parentPath) {
+        // And the parent field to the parent path. Pop from the stack after normalizing the children
+        final String parentField = nestedField.getFieldName();
+        parentPath.addLast(parentField);
+        List<List<QueryComponent>> dnf = toDNF(nestedField.getChild(), negate, parentPath);
+        parentPath.removeLast();
+        return dnf;
     }
 
     private void removeDuplicateConditions(final List<List<QueryComponent>> orOfAnd) {